+++ findContact/api.js
@@ -1,0 +1,2 @@
+export const CATEGORIES_API = 'organizations-storage/categories';
+export const CONTACTS_API = 'organizations-storage/contacts';