+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## (IN PROGRESS)
 
+## [2.0.0](https://github.com/folio-org/ui-plugin-find-contact/tree/v2.0.0) (2020-03-13)
+[Full Changelog](https://github.com/folio-org/ui-plugin-find-contact/compare/v1.4.0...v2.0.0)
+
+* bump the @folio/stripes peer to v3.0.0
+
 ### Stories
 * [UIORGS-112](https://issues.folio.org/browse/UIORGS-112) Sorting when adding contacts does not work
+
+### Bug Fixes
+* [UIPFCONT-3](https://issues.folio.org/browse/UIPFCONT-3) Security update eslint to >= 6.2.1 or eslint-util >= 1.4.1
 
 ## [1.4.0](https://github.com/folio-org/ui-plugin-find-contact/tree/v1.4.0) (2019-12-04)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-contact/compare/v1.3.0...v1.4.0)